requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Pygments==2.17.2
whitenoise==6.6.0